pyproject.toml
[project]
name = "readme-weaver"
version = "1.0.0"
description = "Embed raw markdown partials into README files"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "Oleksandr Ponomarov ", email = "ponomarov.aleksandr@gmail.com" }
]
license = { text = "Apache 2.0" }
dependencies = [
  "loguru>=0.7.2",
  "typer-slim>=0.13.0"
]

[project.scripts]
readme-weaver = "readme_weaver.main:app"

[project.optional-dependencies]
all = [
  "pytest>=8.3.3",
  "ruff>=0.7.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 95
lint.select = ["E", "F", "I", "BLE", "UP"]
lint.ignore = ["UP006", "E501", "E741", "UP035"]
target-version = "py39"
